Add a hook to share across pages. Need to add to others
