+++ food-for-thought/app/home.tsx
@@ -3,7 +3,6 @@ import { router } from "expo-router";
 import { Card, Text, Icon } from '@rneui/themed';
 import SearchBar from "@/components/SearchBar";
 import { TouchableOpacity } from "react-native-gesture-handler";
-import Header from "@/components/Header";
 import RecommendedRestaurant from "@/components/RecommendedRestaurant";
 import MapView, { Marker } from "react-native-maps";
 import { styles } from '../styles/app-styles'; 
@@ -90,7 +89,6 @@ const Home = () => {
       <Card containerStyle={styles.recentCard}>
         <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between' }}>
           <Text h4 style={{ color: '#2E1C47' }}>Last Visited</Text>
-          <Icon name="arrowright" type="antdesign" size={25} onPress={() => console.log("Recent arrow clicked")} />
         </View>
         {/* Rendering the fetched restaurant data in carousel */}
         <FlatList
@@ -106,7 +104,6 @@ const Home = () => {
       <Card containerStyle={styles.recommendationsCard}>
         <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between' }}>
           <Text h4 style={{ color: '#2E1C47' }}>Recommendations</Text>
-          <Icon name="arrowright" type="antdesign" size={25} onPress={() => console.log("Recommendations arrow clicked")} />
         </View>
         {/* passing in static restaurants, but can handle any passed in */}
         <RecommendedRestaurant restaurant={fetchedRestaurants[2]}/>
